requirements.txt
mysql-connector-python
pandas
scikit-learn
faker
matplotlib